refactor structure and add age calculator
all components have been moved to the components directory
and a widged to calculate birth year based on age has been added
