+++ src/App.tsx
@@ -1,7 +1,8 @@
-import Counter from './Counter'
-import ColorButton from './ColorButton'
-import Title from './Title'
-import UserCard from './UserCard'
+import Counter from './components/Counter'
+import ColorButton from './components/ColorButton'
+import Title from './components/Title'
+import UserCard from './components/UserCard'
+import AgeCalc from './components/AgeCalc'
 import './App.css'
 
 function App() {
@@ -14,6 +15,7 @@ function App() {
       <ColorButton />
       <UserCard user={{name: "karel", age:12, email: "bruh@email.com", isStudent: true}}/>
       <UserCard user={{name: "karla", age:85, email: "amogus@email.com"}}/>
+      <AgeCalc />
     </>
   )
 }
