+++ src/utils.ts
@@ -1,0 +1,5 @@
+export function calcAge(age) {
+  const now = new Date().getFullYear()
+  
+  return now - age
+}
